requirements.txt
pandas>=1.5.0
requests>=2.28.0
sqlalchemy>=1.4.0
scikit-learn>=1.1.0
mysql-connector-python>=8.0.0
flask>=2.2.0
openpyxl>=3.0.0
numpy>=1.21.0